运行时数据交互

+++ spaas-workflow-engine-persistent/src/main/java/com/workflow/engine/process/domain/eo/WfProcdefInfo.java
@@ -1,15 +1,10 @@
 package com.workflow.engine.process.domain.eo;
 
 
-import java.util.Date;
-import com.baomidou.mybatisplus.annotation.EnumValue;
 import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
-import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Data;
-
-import java.util.Collection;
 //import io.swagger.annotations.ApiModel;
 //import io.swagger.annotations.ApiModelProperty;
 
@@ -24,9 +19,9 @@ public class WfProcdefInfo{
 
     @TableId
     @TableField(value = "`id`")
-    private Integer  id;
+    private Long  id;
     @TableField(value = "`proc_def_id`")
-    private Integer  procDefId;
+    private Long  procDefId;
     @TableField(value = "`rev`")
     private Integer  rev;
     @TableField(value = "`info_json_id`")
