+++ spaas-workflow-engine-persistent/src/main/java/com/workflow/engine/process/domain/eo/WfRuntimeEventSubscr.java
@@ -1,15 +1,13 @@
 package com.workflow.engine.process.domain.eo;
 
 
-import java.util.Date;
-import com.baomidou.mybatisplus.annotation.EnumValue;
 import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Data;
 
-import java.util.Collection;
+import java.util.Date;
 //import io.swagger.annotations.ApiModel;
 //import io.swagger.annotations.ApiModelProperty;
 
@@ -24,7 +22,7 @@ public class WfRuntimeEventSubscr{
 
     @TableId
     @TableField(value = "`id`")
-    private Integer  id;
+    private Long  id;
     @TableField(value = "`rev`")
     private Integer  rev;
     @TableField(value = "`event_type`")
@@ -32,11 +30,11 @@ public class WfRuntimeEventSubscr{
     @TableField(value = "`event_name`")
     private String eventName;
     @TableField(value = "`execution_id`")
-    private Integer  executionId;
+    private Long  executionId;
     @TableField(value = "`proc_inst_id`")
-    private Integer  procInstId;
+    private Long  procInstId;
     @TableField(value = "`wfivity_id`")
-    private Integer  wfivityId;
+    private Long  wfivityId;
     @TableField(value = "`configuration`")
     private String configuration;
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
