+++ spaas-workflow-engine-persistent/src/main/java/com/workflow/engine/process/domain/eo/WfRuntimeHistoryJob.java
@@ -1,12 +1,13 @@
 package com.workflow.engine.process.domain.eo;
 
 
-import java.util.Date;
 import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Data;
+
+import java.util.Date;
 
 
 /**
@@ -19,7 +20,7 @@ public class WfRuntimeHistoryJob{
 
     @TableId
     @TableField(value = "`id`")
-    private Integer  id;
+    private Long  id;
     @TableField(value = "`rev`")
     private Integer  rev;
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
