+++ spaas-workflow-engine-persistent/src/main/java/com/workflow/engine/process/domain/eo/WfRuntimeTask.java
@@ -1,15 +1,13 @@
 package com.workflow.engine.process.domain.eo;
 
 
-import java.util.Date;
-import com.baomidou.mybatisplus.annotation.EnumValue;
 import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Data;
 
-import java.util.Collection;
+import java.util.Date;
 //import io.swagger.annotations.ApiModel;
 //import io.swagger.annotations.ApiModelProperty;
 
@@ -24,17 +22,17 @@ public class WfRuntimeTask{
 
     @TableId
     @TableField(value = "`id`")
-    private Integer  id;
+    private Long  id;
     @TableField(value = "`rev`")
     private Integer  rev;
     @TableField(value = "`execution_id`")
-    private Integer  executionId;
+    private Long  executionId;
     @TableField(value = "`proc_inst_id`")
-    private Integer  procInstId;
+    private Long  procInstId;
     @TableField(value = "`proc_def_id`")
-    private Integer  procDefId;
+    private Long  procDefId;
     @TableField(value = "`task_def_id`")
-    private Integer  taskDefId;
+    private Long  taskDefId;
     @TableField(value = "`scope_id`")
     private String scopeId;
     @TableField(value = "`sub_scope_id`")
