+++ spaas-workflow-engine-persistent/src/main/java/com/workflow/engine/process/domain/eo/WfRuntimeWfinst.java
@@ -1,15 +1,13 @@
 package com.workflow.engine.process.domain.eo;
 
 
-import java.util.Date;
-import com.baomidou.mybatisplus.annotation.EnumValue;
 import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Data;
 
-import java.util.Collection;
+import java.util.Date;
 //import io.swagger.annotations.ApiModel;
 //import io.swagger.annotations.ApiModelProperty;
 
@@ -24,19 +22,19 @@ public class WfRuntimeWfinst{
 
     @TableId
     @TableField(value = "`id`")
-    private Integer  id;
+    private Long  id;
     @TableField(value = "`rev`")
     private Integer  rev;
     @TableField(value = "`proc_def_id`")
-    private Integer  procDefId;
+    private Long  procDefId;
     @TableField(value = "`proc_inst_id`")
-    private Integer  procInstId;
+    private Long  procInstId;
     @TableField(value = "`execution_id`")
-    private Integer  executionId;
+    private Long  executionId;
     @TableField(value = "`wf_id`")
-    private Integer  wfId;
+    private Long  wfId;
     @TableField(value = "`task_id`")
-    private Integer  taskId;
+    private Long  taskId;
     @TableField(value = "`call_proc_inst_id`")
     private Integer  callProcInstId;
     @TableField(value = "`wf_name`")
